requirements.txt
streamlit==1.37.1
pyocr==0.8.5
pillow==10.4.0
platformdirs==4.2.2
